Refactoring code
Refactoring to use both in CLI and WebUi using Gradio.
Moved many functions into helper.py.
